+++ openedx_events/enterprise/signals.py
@@ -8,16 +8,16 @@ They also must comply with the payload definition specified in
 docs/decisions/0003-events-payload.rst
 """
 
-from openedx_events.enterprise.data import TrackingEvent
+from openedx_events.enterprise.data import LicenseLifecycle
 from openedx_events.tooling import OpenEdxPublicSignal
 
 # .. event_type: org.openedx.enterprise.subscription.license.modified.v1
-# .. event_name: LICENSE_MODIFIED
-# .. event_description: mitted when a subscriptions.License record's data is modified.
-# .. event_data: TrackingEvent
-LICENSE_MODIFIED = OpenEdxPublicSignal(
-    event_type="org.openedx.enterprise.subscription.license.modified.v1",
+# .. event_name: LICENSE_CREATED
+# .. event_description: emitted with Subscription.License is created.
+# .. event_data: LicenseLifecycle
+LICENSE_CREATED = OpenEdxPublicSignal(
+    event_type="org.openedx.enterprise.subscription.license.created.v1",
     data={
-        "user": TrackingEvent,
+        "lifecycle": LicenseLifecycle,
     }
 )